auth/requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
psycopg2==2.9.5
psycopg2-binary==2.9.5  
PyJWT==2.3.0
passlib==1.7.4  # For password hashing utilities
bcrypt==3.2.0  # For strong password hashing
